flask_keras_2_2_4_cpu/requirements.txt
flask==1.0.2
Pillow==6.0.0
tensorflow==1.9.0
Keras==2.2.4